+++ DesignerMacros/src/services-rest-route-validation/command/command-rest-route-validation.ts
@@ -1,0 +1,11 @@
+/// <reference path="../common/common-rest-route-validation.ts" />
+
+/**
+ * Used by Intent.Modules\Modules\Intent.Modules.Metadata.WebApi
+ *
+ * Source code here:
+ * https://github.com/IntentArchitect/Intent.Modules/blob/master/DesignerMacros/src/services-rest-route-validation/command/command-rest-route-validation.ts
+ */
+
+validateRestRoutes(element);
+//return validateRestRoutes(lookup(id));